feat : CalculatorLottoResult 도메인 구현
